Refactor MustSupportTest and MSChecker to delegate extension presence checks to a new Extensions module, improving code organization and maintainability.
