+++ lib/inferno_suite_generator/test_modules/must_support_test.rb
@@ -91,27 +91,9 @@ module InfernoSuiteGenerator
       end
     end
 
-    def must_support_extensions(metadata = nil)
-      prepare_uscdi_ms(:extensions, metadata)
-    end
-
-    def extension_present?(resource, extension_definition)
-      resource_extensions_url_arr = Helpers.extract_extensions_from_resource(resource).map { |ext| ext["url"] }
-      resource_extensions_url_arr.include? extension_definition[:url]
-    end
-
-    def extensions_present_statuses(metadata = nil, resources = [])
-      must_support_extensions(metadata).map do |extension_definition|
-        {
-          definition: extension_definition,
-          url: extension_definition[:url],
-          present: resources.any? { |resource| extension_present?(resource, extension_definition) }
-        }
-      end
-    end
-
     def miss_extensions(metadata = nil, resources = [])
-      found_extensions = extensions_present_statuses(metadata, resources).select do |extension_status|
+      ms_checker = MSChecker.new(metadata, {"exclude_uscdi_only_test": config.options[:exclude_uscdi_only_test]})
+      found_extensions = ms_checker.extensions_present_statuses(resources).select do |extension_status|
         !extension_status[:present]
       end
       found_extensions.map { |extension_status| extension_status[:definition] }
